requirements.txt
# Django Web Framework
Django>=4.2.0,<5.0.0

# Django REST Framework
djangorestframework>=3.14.0,<4.0.0

# JWT Authentication
djangorestframework-simplejwt>=5.2.0,<6.0.0

# CORS Headers (for React frontend)
django-cors-headers>=4.0.0,<5.0.0

# Environment Configuration
python-decouple>=3.8.0,<4.0.0

# PostgreSQL Database Driver
psycopg2-binary>=2.9.0,<3.0.0

# Image Processing (for file uploads)
Pillow>=10.0.0,<12.0.0

# API Documentation (Swagger/OpenAPI)
drf-yasg>=1.21.0,<2.0.0

# Caching (Redis support - optional, falls back to in-memory if not installed)
# Uncomment the next line if using Redis:
# django-redis>=5.2.0,<6.0.0

# Cloudinary Storage (for media files)
cloudinary>=1.36.0,<2.0.0
django-cloudinary-storage>=0.3.0,<1.0.0

# Database URL parser (for Neon PostgreSQL connection)
dj-database-url>=2.1.0,<3.0.0

# WSGI Server (for Render deployment)
gunicorn>=21.2.0,<22.0.0

